# download definitions while generating to speed up plan and update