--- a/lib/kennel/syncer.rb
+++ b/lib/kennel/syncer.rb
@@ -10,12 +10,13 @@ module Kennel
     Plan = Struct.new(:changes, keyword_init: true)
     Change = Struct.new(:type, :api_resource, :tracking_id, :id)
 
-    def initialize(api, expected, kennel:, project_filter: nil, tracking_id_filter: nil)
+    def initialize(api, expected, actual, kennel:, project_filter: nil, tracking_id_filter: nil)
       @api = api
       @kennel = kennel
       @project_filter = project_filter
       @tracking_id_filter = tracking_id_filter
       @expected = Set.new expected # need set to speed up deletion
+      @actual = actual
       calculate_diff
       validate_plan
       prevent_irreversible_partial_updates
@@ -53,7 +54,7 @@ module Kennel
         message = "#{e.class.api_resource} #{e.tracking_id}"
         Kennel.out.puts "Creating #{message}"
         reply = @api.create e.class.api_resource, e.as_json
-        cache_metadata reply, e.class
+        Utils.inline_resource_metadata reply, e.class
         id = reply.fetch(:id)
         changes << Change.new(:create, e.class.api_resource, e.tracking_id, id)
         populate_id_map [], [reply] # allow resolving ids we could previously no resolve
@@ -125,17 +126,15 @@ module Kennel
       @delete = []
       @id_map = IdMap.new
 
-      actual = Progress.progress("Downloading definitions") { download_definitions }
-
       Progress.progress "Diffing" do
-        populate_id_map @expected, actual
-        filter_actual! actual
+        populate_id_map @expected, @actual
+        filter_actual! @actual
         resolve_linked_tracking_ids! @expected # resolve dependencies to avoid diff
 
         @expected.each(&:add_tracking_id) # avoid diff with actual
 
         lookup_map = matching_expected_lookup_map
-        items = actual.map do |a|
+        items = @actual.map do |a|
           e = matching_expected(a, lookup_map)
           if e && @expected.delete?(e)
             [e, a]
@@ -166,19 +165,6 @@ module Kennel
         @delete.sort_by! { |_, _, a| DELETE_ORDER.index a.fetch(:klass).api_resource }
         @update.sort_by! { |_, e, _| DELETE_ORDER.index e.class.api_resource } # slo needs to come before slo alert
       end
-    end
-
-    def download_definitions
-      Utils.parallel(Models::Record.subclasses) do |klass|
-        results = @api.list(klass.api_resource, with_downtimes: false) # lookup monitors without adding unnecessary downtime information
-        results = results[results.keys.first] if results.is_a?(Hash) # dashboards are nested in {dashboards: []}
-        results.each { |a| cache_metadata(a, klass) }
-      end.flatten(1)
-    end
-
-    def cache_metadata(a, klass)
-      a[:klass] = klass
-      a[:tracking_id] = a.fetch(:klass).parse_tracking_id(a)
     end
 
     def ensure_all_ids_found