--- a/lib/kennel/tasks.rb
+++ b/lib/kennel/tasks.rb
@@ -102,8 +102,9 @@ namespace :kennel do
 
   # also generate parts so users see and commit updated generated automatically
   desc "show planned datadog changes (scope with PROJECT=name)"
-  task plan: :generate do
+  task plan: :environment do
     Kennel::Tasks.kennel.plan
+    Kennel::Tasks.kennel.generate
   end
 
   desc "update datadog (scope with PROJECT=name)"