--- a/lib/kennel/utils.rb
+++ b/lib/kennel/utils.rb
@@ -160,6 +160,11 @@ module Kennel
         end
         string
       end
+
+      def inline_resource_metadata(resource, klass)
+        resource[:klass] = klass
+        resource[:tracking_id] = klass.parse_tracking_id(resource)
+      end
     end
   end
 end