--- a/test/kennel/utils_test.rb
+++ b/test/kennel/utils_test.rb
@@ -275,4 +275,13 @@ describe Kennel::Utils do
       Kennel::Utils.pretty_inspect([{ foo: { bar: "bar" } }]).must_equal "[{ foo: { bar: \"bar\" } }]"
     end
   end
+
+  describe ".inline_resource_metadata" do
+    it "adds klass and tracking_id" do
+      resource = { message: "-- Managed by kennel a:b" }
+      Kennel::Utils.inline_resource_metadata(resource, Kennel::Models::Monitor)
+      resource[:tracking_id].must_equal "a:b"
+      resource[:klass].must_equal Kennel::Models::Monitor
+    end
+  end
 end