before (latest release):
```
time bundle exec rake kennel:plan PROJECT=foo
Generating ... 1.51s
Storing ... 0.01s
Downloading definitions ... 7.09s
Diffing ... 0.3s
Plan:
Nothing to do

real	0m10.006s
user	0m3.593s
sys	0m0.551s
```

after:
```
time bundle exec rake kennel:plan PROJECT=foo
Downloading definitions ...
Finding parts ...
Finding parts ... 1.56s
Building json ... 0.01s
Downloading definitions ... 6.6s
Diffing ... 0.1s
Plan:
Nothing to do

real	0m7.228s
user	0m3.455s
sys	0m0.552s
```

rework of https://github.com/grosser/kennel/pull/216

@zdrve