@zdrve 2022-10-30
Doesn't give any advantage if you `rake generate plan` though, does it?