@grosser 2022-10-30
why would anyone `generate plan` when plan already generates ?