@zdrve 2022-10-30
I don't know what to make of this PR yet. The description is misleading, and the tests don't pass. Ping me again when it's ready.