@zdrve 2022-10-30
It's also unclear to me why you felt the need to rework my PR. Yes I know mine isn't ready, but you seem to have taken a whole different approach, I'd love to hear the rationale.