@grosser 2022-10-30
- I made a fresh PR because I though there might be a simpler way of achieving the same goal
- how is the description misleading ?
- tests don't pass in https://github.com/grosser/kennel/pull/216 either and passing tests should not be needed to evaluate the complexity/merit of a PRs approach, afaik this is ready and just needs test fixup unless you can point out any flaws in the approach in which case I might rework or scrap it
- this PR moves the download code and runs it in parallel, to me this seems much simpler than "extract new class, manage threads to know the state of that class, check on the class in multiple places"