@grosser 2022-11-05
fixed up the tests