@zdrve 2022-11-09
Because of the impact it has on the end user (change of behaviour of "plan", and the need to run tasks in the right order), I'm not yet convinced that this approach is simpler. But I get that it's subjective.