@grosser 2022-11-10
There is no change to `rake plan`.
No ordering is required.
`before` is not from the "merge-base", but "latest release", but the time is the same.