@zdrve 2022-11-10
> before is not from the "merge-base"

OK, can you please fix the PR description then?