@zdrve 2022-11-10
> There is no change to rake plan.

Yes there is. `rake plan` used to update `./generated`; now it doesn't.