@grosser 2022-11-10
- aaah I mixed up `generate` and `generated` 🤦
- updated `before` description
- added `generate` back into `plan` so everything should behave as before

idk how widespread `generate plan` usage is (afaik it's not documented anywhere), but we can add a warning or error for that to discourage it if